srcs/containers/services/auth/requirements.txt
flask == 3.1.2
Flask-SQLAlchemy == 3.1.1
marshmallow-sqlalchemy == 1.4.2
python-dotenv == 1.2.1
requests == 2.32.3
mysqlclient == 2.2.7
watchdog
